server/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
bidict==0.21.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Faker==8.11.0
Flask==2.0.1
Flask-Cors==3.0.10
flask-mailman==0.3.0
Flask-Migrate==3.1.0
flask-redis==0.4.0
flask-restx==0.5.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
ghp-import==2.0.1
greenlet==1.1.1
h11==0.12.0
http-ece==1.1.0
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
mkdocs-material==7.2.4
mkdocs-material-extensions==1.0.1
packaging==21.0
psycopg2-binary==2.9.1
py-vapid==1.8.2
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pymdown-extensions==8.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
pywebpush==1.14.0
PyYAML==5.4.1
pyyaml_env_tag==0.1
redis==3.5.3
requests==2.26.0
simple-websocket==0.3.0
six==1.16.0
SQLAlchemy==1.4.22
text-unidecode==1.3
urllib3==1.26.6
watchdog==2.1.3
Werkzeug==2.0.1
wsproto==1.0.0
zipp==3.5.0
